feat: move Namespaces class to lib
