+++ src/kubernetes_cleanup_namespaces/kubernetes_cleanup_namespaces/namespaces.py
@@ -1,32 +1,12 @@
 import logging
 
-from dateutil.parser import parse as parsedatetime
-
 import kubernetes_cleanup_namespaces.context
 from lib.date import older_than_ndays
+from lib.k8s_namespaces import Namespaces
 from lib.kctl import Kctl
 from lib.util import list_subtract
 
 from kubernetes_cleanup_namespaces.config import config
-
-class Namespaces():
-  ''' manage namespaces data '''
-  def __init__(self, kctl):
-    # load namespaces data using the given kctl client
-    self._ns_data = kctl.get_namespaces()
-
-  def created_at(self, namespace_name):
-    ''' given the name of a namespace, return its creation time '''
-    for ns in self._ns_data:
-      if ns['metadata']['name'] == namespace_name:
-        timestamp = ns['metadata']['creationTimestamp']
-        return parsedatetime(timestamp)
-
-  def names(self):
-    ''' return names of namespaces '''
-    return [
-      ns['metadata']['name'] for ns in self._ns_data
-    ]
 
 def cleanup_namespaces():
   ''' delete unprotected namespaces older than n days '''
